cleaned up repo to make more modular
